+++ test/unit/forge/ee/routes/teamBroker/index_spec.js
@@ -184,220 +184,4 @@ describe('Team Broker API', function () {
             response.statusCode.should.equal(404)
         })
     })
-    describe('Test MQTT Broker user auth', function () {
-        before(async function () {
-            await app.inject({
-                method: 'POST',
-                url: `/api/v1/teams/${app.team.hashid}/broker/user`,
-                cookies: { sid: TestObjects.tokens.alice },
-                body: {
-                    username: 'alice',
-                    password: 'aaPassword',
-                    acls: [
-                        {
-                            pattern: 'foo/#',
-                            action: 'both'
-                        }
-                    ]
-                }
-            })
-        })
-        after(async function () {
-            await app.inject({
-                method: 'DELETE',
-                url: `/api/v1/teams/${app.team.hashid}/broker/user/alice`,
-                cookies: { sid: TestObjects.tokens.alice }
-            })
-        })
-        it('Test Authentication pass', async function () {
-            const response = await app.inject({
-                method: 'POST',
-                url: '/api/v1/broker/auth',
-                cookies: { sid: TestObjects.tokens.alice },
-                body: {
-                    username: `alice@${app.team.hashid}`,
-                    password: 'aaPassword',
-                    clientId: `alice@${app.team.hashid}`
-                }
-            })
-            response.statusCode.should.equal(200)
-            const result = response.json()
-            result.should.have.property('result', 'allow')
-            result.should.have.property('is_superuser', false)
-            result.should.have.property('client_attrs')
-            result.client_attrs.should.have.property('team')
-        })
-        it('Test Authentication fail', async function () {
-            const response = await app.inject({
-                method: 'POST',
-                url: '/api/v1/broker/auth',
-                cookies: { sid: TestObjects.tokens.alice },
-                body: {
-                    username: `alice@${app.team.hashid}`,
-                    password: 'bbPassword',
-                    clientId: `alice@${app.team.hashid}`
-                }
-            })
-            response.statusCode.should.equal(200)
-            const result = response.json()
-            result.should.have.property('result', 'deny')
-        })
-        it('Test Authentication fail none existent user', async function () {
-            const response = await app.inject({
-                method: 'POST',
-                url: '/api/v1/broker/auth',
-                cookies: { sid: TestObjects.tokens.alice },
-                body: {
-                    username: `alice@${app.team.hashid}-foo`,
-                    password: 'bbPassword',
-                    clientId: `alice@${app.team.hashid}`
-                }
-            })
-            response.statusCode.should.equal(200)
-            const result = response.json()
-            result.should.have.property('result', 'deny')
-        })
-        it('Test Authentication fail no password', async function () {
-            const response = await app.inject({
-                method: 'POST',
-                url: '/api/v1/broker/auth',
-                cookies: { sid: TestObjects.tokens.alice },
-                body: {
-                    username: `alice@${app.team.hashid}`,
-                    password: '',
-                    clientId: `alice@${app.team.hashid}`
-                }
-            })
-            response.statusCode.should.equal(200)
-            const result = response.json()
-            result.should.have.property('result', 'deny')
-        })
-        it('Test subscribe allowed', async function () {
-            const response = await app.inject({
-                method: 'POST',
-                url: '/api/v1/broker/acls',
-                cookies: { sid: TestObjects.tokens.alice },
-                body: {
-                    username: `alice@${app.team.hashid}`,
-                    topic: 'foo/bar',
-                    action: 'subscribe'
-                }
-            })
-            response.statusCode.should.equal(200)
-            const result = response.json()
-            result.should.have.property('result', 'allow')
-        })
-        it('Test subscribe not allowed', async function () {
-            const response = await app.inject({
-                method: 'POST',
-                url: '/api/v1/broker/acls',
-                cookies: { sid: TestObjects.tokens.alice },
-                body: {
-                    username: `alice@${app.team.hashid}`,
-                    topic: 'bar/foo',
-                    action: 'subscribe'
-                }
-            })
-            response.statusCode.should.equal(200)
-            const result = response.json()
-            result.should.have.property('result', 'deny')
-        })
-        it('Test publish allowed', async function () {
-            const response = await app.inject({
-                method: 'POST',
-                url: '/api/v1/broker/acls',
-                cookies: { sid: TestObjects.tokens.alice },
-                body: {
-                    username: `alice@${app.team.hashid}`,
-                    topic: 'foo/foo',
-                    action: 'publish'
-                }
-            })
-            response.statusCode.should.equal(200)
-            const result = response.json()
-            result.should.have.property('result', 'allow')
-        })
-        it('Test publish not allowed', async function () {
-            const response = await app.inject({
-                method: 'POST',
-                url: '/api/v1/broker/acls',
-                cookies: { sid: TestObjects.tokens.alice },
-                body: {
-                    username: `alice@${app.team.hashid}`,
-                    topic: 'bar/foo',
-                    action: 'publish'
-                }
-            })
-            response.statusCode.should.equal(200)
-            const result = response.json()
-            result.should.have.property('result', 'deny')
-        })
-        /*
-         * Need tests for the project nodes and devices both
-         * Auth and ACL
-         */
-        it('Test Authentication forge_platform pass', async function () {
-            const response = await app.inject({
-                method: 'POST',
-                url: '/api/v1/broker/auth',
-                cookies: { sid: TestObjects.tokens.alice },
-                body: {
-                    username: 'forge_platform',
-                    password: await app.settings.get('commsToken'),
-                    clientId: 'alice'
-                }
-            })
-            response.statusCode.should.equal(200)
-            const result = response.json()
-            result.should.have.property('result', 'allow')
-            result.should.have.property('is_superuser', false)
-            result.should.have.property('client_attrs')
-            result.client_attrs.should.have.property('team', 'team/internal/')
-        })
-        it('Test Authentication forge_platform fail', async function () {
-            const response = await app.inject({
-                method: 'POST',
-                url: '/api/v1/broker/auth',
-                cookies: { sid: TestObjects.tokens.alice },
-                body: {
-                    username: 'forge_platform',
-                    password: 'fooo',
-                    clientId: 'forge_platform'
-                }
-            })
-            response.statusCode.should.equal(200)
-            const result = response.json()
-            result.should.have.property('result', 'deny')
-        })
-        it('Test Authorization forge_platform subscribe', async function () {
-            const response = await app.inject({
-                method: 'POST',
-                url: '/api/v1/broker/acls',
-                cookies: { sid: TestObjects.tokens.alice },
-                body: {
-                    username: 'forge_platform',
-                    topic: 'ff/v1/+/l/+/status',
-                    action: 'subscribe'
-                }
-            })
-            response.statusCode.should.equal(200)
-            const result = response.json()
-            result.should.have.property('result', 'allow')
-        })
-        it('Test Authorization forge_platform publish deny', async function () {
-            const response = await app.inject({
-                method: 'POST',
-                url: '/api/v1/broker/acls',
-                cookies: { sid: TestObjects.tokens.alice },
-                body: {
-                    username: 'forge_platform',
-                    topic: 'ff/v1/+/l/+/status',
-                    action: 'publish'
-                }
-            })
-            response.statusCode.should.equal(200)
-            const result = response.json()
-            result.should.have.property('result', 'deny')
-        })
-    })
 })
